Convert RegisterCafe modal to a screen.
