+++ App/Navigation/AppNavigation/Modal/Settings/index.ts
@@ -3,8 +3,8 @@ import { createStackNavigator } from 'react-navigation'
 import Account from '../../../../SB/views/UserProfile'
 import NotificationSettings from '../../../../SB/views/Notifications'
 import Cafes from '../../../../Containers/Cafes'
+import RegisterCafe from '../../../../Containers/RegisterCafe'
 import Cafe from '../../../../Containers/Cafe'
-import CafeSession from '../../../../screens/cafe-session'
 import Storage from '../../../../SB/views/Storage'
 import DeviceLogs from '../../../../SB/views/DeviceLogs'
 import NodeLogsScreen from '../../../../Components/NodeLogsScreen'
@@ -18,8 +18,8 @@ const nav = createStackNavigator(
     Account,
     NotificationSettings,
     Cafes,
+    RegisterCafe,
     Cafe,
-    CafeSession,
     Storage,
     RecoveryPhrase,
     ChangeAvatar: SetAvatar,
